% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunks_to_process.R
\name{build_sh_file}
\alias{build_sh_file}
\title{make an sh file for the HPC from a list of catchments and chunks}
\usage{
build_sh_file(
  misslist,
  summaryFun,
  runImmediate = FALSE,
  lengthOrChunk = c("short", "long")
)
}
\arguments{
\item{misslist}{list with names being catchments, and each having a numeric vector of chunks to run}

\item{summaryFun}{function for the summary}

\item{runImmediate}{logical FALSE- immediately start the sbatch}

\item{lengthOrChunk}{'short', 'long' or 'chunk'. Short and long are times, 'chunk' makes subchunks, and is rarely used or maintained.}
}
\value{
nothing, makes a file
}
\description{
make an sh file for the HPC from a list of catchments and chunks
}
