% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{catchAggW}
\alias{catchAggW}
\title{Spatial aggregation of stars objects}
\usage{
catchAggW(starObj, weights, FUN, summaryPoly)
}
\arguments{
\item{starObj}{stars object to aggregate}

\item{weights}{a vector of the weights to give each stricture (typically area of relevant polygons)}

\item{FUN}{aggregation function}

\item{summaryPoly}{summary polygon(s) to aggregate into (sf)}
}
\description{
Spatial aggregation of stars objects
}
