% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chunks_to_process.R
\name{chunks_to_process}
\alias{chunks_to_process}
\title{Find the chunks that we need to process}
\usage{
chunks_to_process(
  out_dir,
  dataname,
  summaryFun,
  catchment = "all",
  nchunks = NULL,
  poly_per_chunk = NULL,
  extraname = NULL,
  returnForR = TRUE,
  produce_sh = FALSE,
  filetype = ".rds"
)
}
\arguments{
\item{out_dir}{outer directory for all output data}

\item{dataname}{name of the dataset (this will be the next dir in)}

\item{summaryFun}{name of the summary function (the next dir in, enabling multiple summaries of same data)}

\item{catchment}{names of catchments or other units. Default 'all' forces all catchments in the basin.}

\item{nchunks}{number of chunks. Default NULL. If > 1, data goes into a 'catchment/chunked/...' directory. If 1, does not chunk, data goes into the summaryFun directory}

\item{poly_per_chunk}{number of polygons per chunk. default NULL. Dynamically adjusts the number of chunks per catchment so each has approximately polys_per_chunk. Cannot use this and nchunks.}

\item{returnForR}{logical, default TRUE, return a tibble with catchment and chunk columns}

\item{produce_sh}{logical default FALSE, produce an .sh file if TRUE}

\item{filetype}{'.rds' (default), or '.rdata' for backwards compatibility}
}
\value{
a tibble or invisible, depending on returnForR
}
\description{
Find the chunks that we need to process
}
