% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{clean_area}
\alias{clean_area}
\title{Make sure areas aren't larger than areas of polygons. There are cases where the rounding is an issue.}
\usage{
clean_area(in_geo, anaes, test = FALSE)
}
\arguments{
\item{in_geo}{input geometry (either stars or sf)}

\item{anaes}{the canonical set of anae polygons}
}
\value{
clean stars or sf object with areas never larger than the anaes (and in the case of ANAEs, the geometry is replaced)
}
\description{
Make sure areas aren't larger than areas of polygons. There are cases where the rounding is an issue.
}
