% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concat_star_index.R
\name{concat_star_index}
\alias{concat_star_index}
\title{concatenate lists of stars objects and indices}
\usage{
concat_star_index(
  starindexlist,
  dimension,
  starname = names(starindexlist[[1]][1]),
  indexname = names(starindexlist[[1]][2])
)
}
\arguments{
\item{starindexlist}{a list of lists with the stars in 1 and indices in 1}

\item{dimension}{dimension along which to \code{\link[=c.stars]{c.stars()}}}

\item{starname}{name to give the stars object. Defaults to the name of the first one}

\item{indexname}{name to give the index, defaults to the name of the first one}
}
\value{
a list with stars in 1 and indices in 2, concatenated together over the starindexlist
}
\description{
Because the list structure means we can't just c() the stars and bind_rows the indices, since they're items 1 and 2 in a list
}
