% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getBimonth}
\alias{getBimonth}
\title{Get the bimonthly group}
\usage{
getBimonth(input.date)
}
\arguments{
\item{input.date}{a date}
}
\value{
the bimonth period a date is in.
}
\description{
Helps match inundation layer
}
