% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{getWaterYear}
\alias{getWaterYear}
\title{get the water year that a date is in}
\usage{
getWaterYear(input.date)
}
\arguments{
\item{input.date}{a date}
}
\value{
the water year the date is in
}
\description{
get the water year that a date is in
}
