% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_chunk_dirs.R
\name{make_chunk_dirs}
\alias{make_chunk_dirs}
\title{Set up directories for chunked (or not) data and stricture processing}
\usage{
make_chunk_dirs(
  out_dir,
  catchment,
  process_function,
  dataname,
  nchunks,
  subchunkArgs = NULL,
  extraname = NULL
)
}
\arguments{
\item{catchment}{catchment/larger group name}

\item{process_function}{the name of the function used to process, same as \code{summaryFun} for data or \code{strict_fun} for strictures. Works best if character so name does not get lost in the call stack}

\item{dataname}{input data name. Should be \code{''} if strictures}

\item{nchunks}{nchunks}

\item{subchunkArgs}{usually ignored, not maintained much}

\item{extraname}{as in other functions}
}
\value{
a list with scriptOut, unique_chunkname and unique_indexname
}
\description{
Set up directories for chunked (or not) data and stricture processing
}
