% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSHfails.R
\name{makeSHfails}
\alias{makeSHfails}
\title{Find missing runs}
\usage{
makeSHfails(
  outerDir,
  varName,
  summaryFuns,
  nchunks = 100,
  lengthOrChunk,
  runImmediate = FALSE,
  forceAllCatchments = FALSE,
  returnForR = FALSE,
  produce_sh = TRUE
)
}
\arguments{
\item{produce_sh}{}
}
\description{
Find missing runs
}
