% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchStarsIndex.R
\name{matchStarsIndex}
\alias{matchStarsIndex}
\title{Function to sort stars according to their index files and deal with any duplicated ANAEs}
\usage{
matchStarsIndex(
  index1,
  stars1 = NULL,
  index2,
  stars2,
  indexcol = c(1, 1),
  testfinal = TRUE,
  return1 = FALSE,
  as_test = FALSE
)
}
\arguments{
\item{index1}{sf with an index in column 1 and geometry}

\item{stars1}{a stars belonging to index1. Default NULL}

\item{index2}{sf with an index in column 1 and geometry}

\item{stars2}{a stars belonging to index2}

\item{indexcol}{columns of the index values. Default c(1,1) if both index1 and 2 have the column as 1}

\item{testfinal}{do a spatial check of the stars itself. This takes a very long time.}

\item{return1}{whether to return index1 or not. Default FALSE}

\item{as_test}{Use this as a test of whether things needed to be sorted (ie emits error if sorting happens). Default FALSE}
}
\value{
a list of stars and index (and if return1, index1 also)
}
\description{
Particularly important while parallel processing. Often used with anaes as index1 to sort stars relative to them. In typical use, this assumes the stars2 and index2 are not shuffled relative to each other, and sorts them relative to index1. If \code{testfinal = TRUE}, it does a check of the stars2 and index2 spatial matching, but that is slow.
}
