% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_funs.R
\name{nafuns}
\alias{nafuns}
\alias{propor}
\alias{sumna}
\alias{meanna}
\alias{maxna}
\alias{minna}
\title{Common functions that handle na.rm = TRUE but return NA instead of 0 when all NA}
\usage{
propor(x, na.rm = TRUE)

sumna(x, na.rm = TRUE)

meanna(x, na.rm = TRUE)

maxna(x, na.rm = TRUE)

minna(x, na.rm = TRUE)
}
\arguments{
\item{x}{vector}

\item{na.rm}{standard na.rm behavior. Default TRUE unlike the base functions}
}
\description{
Common functions that handle na.rm = TRUE but return NA instead of 0 when all NA
}
