% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_data.R
\name{parallel_data}
\alias{parallel_data}
\title{Light parallel wrapper over \code{\link[=processData]{processData()}}}
\usage{
parallel_data(
  runframe,
  dataname,
  data_dir,
  poly_path = file.path(out_dir, "ANAEprocessed"),
  summaryFun,
  out_dir,
  nchunks = 100,
  whichcrs = 3577,
  maxPix = 1e+05,
  rastRollArgs = NULL,
  saveout = TRUE
)
}
\arguments{
\item{runframe}{a tibble or dataframe with columns 'catchment' and 'chunk'. if nchunks == 1, the 'chunk' column still needs to be there, but there should only be one row per catchment and chunk should always be 1}
}
\value{
a tibble of timings. The main output is saving the data to out_dir.
}
\description{
Light parallel wrapper over \code{\link[=processData]{processData()}}
}
