% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_anae.R
\name{process_anae}
\alias{process_anae}
\title{Process the ANAE files including intersecting}
\usage{
process_anae(
  datIn,
  outDir,
  outname = "ANAE",
  intersections = c("climate", "valleys"),
  commonCRS = 3577,
  chunksave = "ValleyName"
)
}
\arguments{
\item{datIn}{path to directory with the ANAE}

\item{outDir}{path to output. Typically 'datOut/ANAEprocessed' or similar}

\item{outname}{name of this particular output, e.g. 'ANAEbasinclim'}

\item{intersections}{character vector or list of characters and sf objects to intersect with the anaes}

\item{commonCRS}{default 3577}

\item{chunksave}{If want to save chunks of the ANAEs, this is the column name to chunk into, e.g. 'ValleyName'. If NULL, only the full ANAE set is saved.}
}
\description{
Process the ANAE files including intersecting
}
