% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/processData.R
\name{process_data}
\alias{process_data}
\title{General function to process rasters into polygons}
\usage{
process_data(
  dataname,
  data_dir,
  poly_path = file.path(out_dir, "ANAEprocessed"),
  summaryFun,
  out_dir,
  catchment,
  thischunk,
  subchunkArgs = NULL,
  nchunks = 100,
  whichcrs = 3577,
  maxPix = 1e+05,
  rastRollArgs = NULL,
  saveout = TRUE
)
}
\arguments{
\item{dataname}{name of the data to process (and name of output directory for the data type). Current options 'inundation', 'soil_temp', 'moisture'}

\item{data_dir}{directory with raw data}

\item{poly_path}{path to the polygons to process into}

\item{summaryFun}{character name of the summary function}

\item{out_dir}{output directory}

\item{catchment}{catchment to chunk into}

\item{thischunk}{helps with chunking for parallel}

\item{subchunkArgs}{more chunking help, default NULL}

\item{nchunks}{default 100, how many chunks to use. If 1, the chunked system cancels out and the data is just saved to out_dir/dataname/summaryFun/Catchment_summaryFun.rds.}

\item{whichcrs}{desired crs}

\item{maxPix}{maximum number of pixels before individual polygons are broken up in \code{\link[=rastPolyJoin]{rastPolyJoin()}}}

\item{rastRollArgs}{arguments to \code{\link[=rastPolyJoin]{rastPolyJoin()}} to roll the raster before aggregating}

\item{saveout}{default TRUE, save data out.}
}
\value{
a tibble of the time taken, main output is the saved data
}
\description{
General function to process rasters into polygons
}
