% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_with_checks.R
\name{process_with_checks}
\alias{process_with_checks}
\title{identify needed analyes, run, concatenate, and check}
\usage{
process_with_checks(
  dataname,
  data_dir,
  poly_path = file.path(out_dir, "ANAEprocessed"),
  summaryFun,
  out_dir,
  catchment = "all",
  nchunks = 100,
  whichcrs = 3577,
  maxPix = 1e+05,
  rastRollArgs = NULL,
  extraname = NULL
)
}
\arguments{
\item{dataname}{name of the data to process (and name of output directory for the data type). Current options 'inundation', 'soil_temp', 'moisture'}

\item{data_dir}{directory with raw data}

\item{poly_path}{path to the polygons to process into}

\item{summaryFun}{character name of the summary function}

\item{out_dir}{output directory}

\item{nchunks}{default 100, how many chunks to use. If 1, the chunked system cancels out and the data is just saved to out_dir/dataname/summaryFun/Catchment_summaryFun.rds.}

\item{whichcrs}{desired crs}

\item{maxPix}{maximum number of pixels before individual polygons are broken up in \code{\link[=rastPolyJoin]{rastPolyJoin()}}}

\item{rastRollArgs}{arguments to \code{\link[=rastPolyJoin]{rastPolyJoin()}} to roll the raster before aggregating}

\item{extraname}{an extra level for the directory path, so, out_dir/dataname/summaryFun/extraname. Often used if there are multiple sets of rastRollArgs that get the same summaryFun, e.g. min moisture over different timeperiods.}
}
\value{
if failures, a tibble, else invisible
}
\description{
wrapper over \code{\link[=chunks_to_process]{chunks_to_process()}}, \code{\link[=parallel_data]{parallel_data()}}, \code{\link[=concat_chunks]{concat_chunks()}}, and
\code{\link[=test_anae_agg]{test_anae_agg()}}, since they are often run together and take the same
arguments.
}
