% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rastPolyJoin.R
\name{rastPolyJoin}
\alias{rastPolyJoin}
\title{Intersect raster and polygons and get spatial average and return stars}
\usage{
rastPolyJoin(
  polysf,
  rastst,
  grouper = "UID",
  FUN = weighted.mean,
  maintainPolys = TRUE,
  na.replace = NA,
  whichcrs = 3577,
  maxPixels = 1e+05,
  pixelsize = NA,
  rastRollArgs = NULL
)
}
\arguments{
\item{polysf}{polygons in sf}

\item{rastst}{rasters in stars or stars_proxy}

\item{grouper}{a single grouping variable to identify single polygons in polysf}

\item{FUN}{the aggregation function to apply. Should be a bare function name
(tricky to get deep in the stack), a character, or a list, e.g. \verb{list(testf = \\(x) sum(ifelse(x > 0.005, area, 0), na.rm = TRUE))}. Due to changes in
dplyr 1.1, if functions need 'area', it should be internally hardcoded, not
a second argument. e.g. the easiest way to use weighted.mean is to write a
new function \code{wm <- function(x) {weighted.mean(x, area, na.rm = T)}} and
pass 'wm'. There are other solutions, but that seems to be the cleanest.}

\item{maintainPolys}{TRUE averages back into them. FALSE leaves them split, and makes a new grouping variable to keep it distinct}

\item{na.replace}{allows replacing NA with numerical values (e.g. 0)}

\item{whichcrs}{Default here is the Australian Albers 3577. If geographic (unprojected) crs is used, the intersection jumbles up for sf version >=1.0, and so I'm forcing a projection as default}

\item{maxPixels}{maximum number of pixels before internally chunking}

\item{pixelsize}{size of the pixels in the raster}

\item{rastRollArgs}{allows passing a list of arguments to timeRoll to roll the  raster after cropping. use the \code{attribute_number} format, e.g. \code{rastRollArgs  = list(attribute_number = 1, tDim = 3, FUN = RcppRoll::roll_max, rolln = 2, align = 'right', na.rm = TRUE)}. NULL (the default) just bypasses}
}
\description{
Intersect raster and polygons and get spatial average and return stars
}
