% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_catchment_polys.R
\name{read_catchment_polys}
\alias{read_catchment_polys}
\title{Helper to read in groups of polygons (typically ANAEs broken into catchments)}
\usage{
read_catchment_polys(poly_path, catchment)
}
\arguments{
\item{poly_path}{path to the polygons to process into}

\item{catchment}{catchment to chunk into}
}
\value{
valid sf of the polys for the requested group \code{catchment}
}
\description{
Helper to read in groups of polygons (typically ANAEs broken into catchments)
}
