% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sf_and_aggforce}
\alias{sf_and_aggforce}
\title{Turn a stars into an sf and aggregate it into a larger poly}
\usage{
sf_and_aggforce(starsobj, catchpoly, newname, funlist)
}
\arguments{
\item{starsobj}{stars object}

\item{catchpoly}{the larger polygon. There is no spatial joining here- it's just a force}

\item{newname}{the data column name}

\item{funlist}{function (bare or \(x) style)}
}
\value{
an sf aggregated over all of the stars polygons and attached to the catchpoly polygon
}
\description{
Turn a stars into an sf and aggregate it into a larger poly
}
