% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sfandcatch}
\alias{sfandcatch}
\title{Make a stars an sf and join with spatial info about its polygons or larger ones}
\usage{
sfandcatch(starsObj, newname, polyinfo = NULL, larger_poly = NULL)
}
\arguments{
\item{starsObj}{stars object}

\item{newname}{name for the column of data}

\item{polyinfo}{info about these polygons (sf). These are st_joined by st_equals_exact}

\item{larger_poly}{info about other polygons (sf). These are just st_joined with left = FALSE}
}
\value{
an sf with geometry as in the original stars and possibly additional info from the other polygons
}
\description{
Make a stars an sf and join with spatial info about its polygons or larger ones
}
