% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{tempaggregate}
\alias{tempaggregate}
\title{Temporal aggregation of stars objects}
\usage{
tempaggregate(starObj, by_t, FUN, na.rm = TRUE, dates_end_interval = FALSE)
}
\arguments{
\item{starObj}{stars}

\item{by_t}{the time intervals, as in \code{\link[=aggregate]{aggregate()}}, can be numeric or things like 'years'}

\item{FUN}{aggregation function}

\item{na.rm}{as usual}

\item{dates_end_interval}{default FALSE, dates are the start of the interval, if TRUE, dates are the end.}
}
\value{
an aggregated stars object
}
\description{
THIS RETURNS TIMES AT THE \emph{START} OF THE INTERVAL unless \code{dates_end_interval = TRUE}
Note that inundation dates are the start of the interval, as are the others. There was confusion about this previously.
}
