% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_anae_agg.R
\name{test_anae_agg}
\alias{test_anae_agg}
\title{Check the UIDs haven't gotten scrambled or duplicated}
\usage{
test_anae_agg(
  catchment,
  out_dir,
  dataname,
  summaryFun,
  poly_path = file.path(out_dir, "ANAEprocessed"),
  debugbreak = FALSE,
  filetype = ".rds"
)
}
\arguments{
\item{debugbreak}{for testing}

\item{filetype}{'.rds' (default), or '.rdata' for backwards compatibility}
}
\value{
tibble of pass fails
}
\description{
Check the UIDs haven't gotten scrambled or duplicated
}
