% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeRoll.R
\name{timeRoll}
\alias{timeRoll}
\title{Time-roll stars}
\usage{
timeRoll(
  stardat,
  attribute_number = 1,
  tDim = NULL,
  FUN,
  rolln = 1L,
  align = "center",
  fill = c(NA, NA, NA),
  na.rm = FALSE
)
}
\arguments{
\item{stardat}{a stars object or a single attribute, e.g. stars_obj[\link{1}]}

\item{attribute_number}{the attribute number if a stars object}

\item{tDim}{the number of the time dimension, default NULL chooses final dimension}

\item{FUN}{function to apply. Typically RcppRoll::roll_*}

\item{rolln}{rolling window. see RcppRoll}

\item{align}{see RcppRoll}

\item{fill}{see RcppRoll}

\item{na.rm}{see RcppRoll}
}
\value{
an array or stars object, depending on what was given.
}
\description{
Time-roll stars
}
