% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unevenTimeMult.R
\name{unevenTimeMult}
\alias{unevenTimeMult}
\title{multiply one stars by another at a different temporal scale, possibly with a lag}
\usage{
unevenTimeMult(fineStars, coarseStars, lag, invertCoarseLogic = FALSE)
}
\arguments{
\item{fineStars}{stars with a fine timestep}

\item{coarseStars}{stars on a coarser timestep}

\item{lag}{lag from fine to coarse; i.e. 0 is multiply day by coarse same year, 1 is day by previous year}

\item{invertCoarseLogic}{flip the logicals on the coarse, ie if they are a competitor species}
}
\value{
multiplied stars with same dimensions as fineStars
}
\description{
ie daily for one species depends on another ever succeeding in the previous year
}
