% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CC2-package.R
\docType{package}
\name{CC2-package}
\alias{CC2}
\alias{CC2-package}
\title{CC2: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Galen Holt \email{g.holt@deakin.edu.au} (\href{https://orcid.org/0000-0002-7455-9275}{ORCID})

Authors:
\itemize{
  \item Ashley Macqueen \email{ashley.macqueen@vewh.vic.gov.au}
}

}
\keyword{internal}
