% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{clean_area}
\alias{clean_area}
\title{Make sure areas aren't larger than areas of polygons. There are cases where the rounding is an issue.}
\usage{
clean_area(starsObj, anaes)
}
\arguments{
\item{anaes}{}
}
\value{
clean starsObj
}
\description{
Make sure areas aren't larger than areas of polygons. There are cases where the rounding is an issue.
}
