% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatANAEchunks.R
\name{concat_chunks}
\alias{concat_chunks}
\title{Function to concatenate chunked ANAE outputs}
\usage{
concat_chunks(
  out_dir,
  dataname,
  summaryFun,
  catchment = "all",
  poly_path = file.path(out_dir, "ANAEprocessed"),
  extraname = NULL,
  rebuild = FALSE,
  filetype = ".rds"
)
}
\arguments{
\item{out_dir}{output directory}

\item{dataname}{name of the data to process (and name of output directory for the data type). Current options 'inundation', 'soil_temp', 'moisture'}

\item{summaryFun}{character name of the summary function}

\item{catchment}{catchment to chunk into}

\item{poly_path}{path to the polygons to process into}

\item{extraname}{an extra level for the directory path, so, out_dir/dataname/summaryFun/extraname. Often used if there are multiple sets of rastRollArgs that get the same summaryFun, e.g. min moisture over different timeperiods.}

\item{rebuild}{logical default FALSE does not re-concatenate if a concatenated file exists. If TRUE, it does.}

\item{filetype}{'.rds' (default), or '.rdata' for backwards compatibility}
}
\value{
invisible, main thing is the saving of concatenated data.
}
\description{
Function to concatenate chunked ANAE outputs
}
