% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatANAEchunks.R
\name{concat_chunks}
\alias{concat_chunks}
\title{Function to concatenate chunked ANAE outputs}
\usage{
concat_chunks(
  out_dir,
  dataname,
  summaryFun,
  catchment = "all",
  rebuild = FALSE,
  filetype = ".rds"
)
}
\arguments{
\item{rebuild}{logical default FALSE does not re-concatenate if a concatenated file exists. If TRUE, it does.}

\item{filetype}{'.rds' (default), or '.rdata' for backwards compatibility}
}
\value{
invisible, main thing is the saving of concatenated data.
}
\description{
Function to concatenate chunked ANAE outputs
}
