% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_with_checks.R
\name{process_with_checks}
\alias{process_with_checks}
\title{identify needed analyes, run, concatenate, and check}
\usage{
process_with_checks(
  dataname,
  data_dir,
  poly_path = file.path(out_dir, "ANAEprocessed"),
  process_function,
  out_dir,
  catchment = "all",
  nchunks = NULL,
  poly_per_chunk = NULL,
  whichcrs = 3577,
  maxPix = 1e+05,
  rastRollArgs = NULL,
  extraname = NULL
)
}
\arguments{
\item{dataname}{as in \code{\link[=parallel_data]{parallel_data()}}, but if \code{''} or missing, assumes strictures, since they have no dataname argument.}

\item{data_dir}{directory with raw data}

\item{poly_path}{path to the polygons to process into}

\item{process_function}{the name of the function used to process, same as \code{summaryFun} for data and \code{strict_fun} for strictures}

\item{out_dir}{output directory}

\item{catchment}{names of catchments or other units. Default 'all' forces all catchments in the basin.}

\item{nchunks}{number of chunks. Default NULL. If > 1, data goes into a 'catchment/chunked/...' directory. If 1, does not chunk, data goes into the summaryFun directory}

\item{poly_per_chunk}{number of polygons per chunk. default NULL. Dynamically adjusts the number of chunks per catchment so each has approximately polys_per_chunk. Cannot use this and nchunks.}

\item{whichcrs}{desired crs}

\item{maxPix}{maximum number of pixels before individual polygons are broken up in \code{\link[=rastPolyJoin]{rastPolyJoin()}}}

\item{rastRollArgs}{arguments to \code{\link[=rastPolyJoin]{rastPolyJoin()}} to roll the raster before aggregating}

\item{extraname}{an extra level for the directory path, so, out_dir/dataname/summaryFun/extraname. Often used if there are multiple sets of rastRollArgs that get the same summaryFun, e.g. min moisture over different timeperiods.}
}
\value{
if failures, a tibble, else invisible
}
\description{
wrapper over \code{\link[=chunks_to_process]{chunks_to_process()}}, \code{\link[=parallel_data]{parallel_data()}}, \code{\link[=concat_chunks]{concat_chunks()}}, and
\code{\link[=test_anae_agg]{test_anae_agg()}}, since they are often run together and take the same
arguments.
}
