% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_anae_agg.R
\name{test_anae_agg}
\alias{test_anae_agg}
\title{Check the UIDs haven't gotten scrambled or duplicated}
\usage{
test_anae_agg(
  catchment,
  out_dir,
  dataname,
  summaryFun,
  poly_path = file.path(out_dir, "ANAEprocessed"),
  extraname = NULL,
  debugbreak = FALSE,
  filetype = ".rds"
)
}
\arguments{
\item{catchment}{catchment to chunk into}

\item{out_dir}{output directory}

\item{dataname}{name of the data to process (and name of output directory for the data type). Current options 'inundation', 'soil_temp', 'moisture'}

\item{summaryFun}{character name of the summary function}

\item{poly_path}{path to the polygons to process into}

\item{extraname}{an extra level for the directory path, so, out_dir/dataname/summaryFun/extraname. Often used if there are multiple sets of rastRollArgs that get the same summaryFun, e.g. min moisture over different timeperiods.}

\item{debugbreak}{for testing}

\item{filetype}{'.rds' (default), or '.rdata' for backwards compatibility}
}
\value{
tibble of pass fails
}
\description{
Check the UIDs haven't gotten scrambled or duplicated
}
